Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
tiktoken==0.5.2 